requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3